presidio-analyzer/presidio_analyzer/predefined_recognizers/requirements.txt
anyio==3.6.2
aptos-sdk==0.6.0
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
asttokens==2.1.0
attrs==22.1.0
backcall==0.2.0
beautifulsoup4==4.11.1
bleach==5.0.1
blis==0.7.9
catalogue==2.0.8
certifi @ file:///croot/certifi_1665076670883/work/certifi
cffi==1.15.1
charset-normalizer==2.1.1
click==8.1.3
comm==0.1.1
confection==0.0.3
cymem==2.0.7
debugpy==1.6.3
decorator==5.1.1
defusedxml==0.7.1
en-core-web-lg @ https://github.com/explosion/spacy-models/releases/download/en_core_web_lg-3.4.1/en_core_web_lg-3.4.1-py3-none-any.whl
entrypoints==0.4
exceptiongroup==1.0.4
executing==1.2.0
Faker==15.3.3
fastjsonschema==2.16.2
filelock==3.8.0
haikunator==2.1.0
idna==3.4
importlib-metadata==5.1.0
iniconfig==1.1.1
ipykernel==6.18.1
ipython==8.7.0
ipython-genutils==0.2.0
ipywidgets==8.0.2
iso3166==2.1.1
jedi==0.18.2
Jinja2==3.1.2
joblib==1.4.2
jsonschema==4.17.1
jupyter==1.0.0
jupyter-console==6.4.4
jupyter-server==1.23.3
jupyter_client==7.4.7
jupyter_core==5.1.0
jupyterlab-pygments==0.2.2
jupyterlab-widgets==3.0.3
ko-core-news-lg @ https://github.com/explosion/spacy-models/releases/download/ko_core_news_lg-3.4.0/ko_core_news_lg-3.4.0-py3-none-any.whl
ko-core-news-sm @ https://github.com/explosion/spacy-models/releases/download/ko_core_news_sm-3.4.0/ko_core_news_sm-3.4.0-py3-none-any.whl
langcodes==3.3.0
MarkupSafe==2.1.1
matplotlib-inline==0.1.6
mistune==2.0.4
murmurhash==1.0.9
nbclassic==0.4.8
nbclient==0.7.0
nbconvert==7.2.5
nbformat==5.7.0
nest-asyncio==1.5.6
notebook==6.5.2
notebook_shim==0.2.2
numpy==1.23.5
packaging==21.3
pandas==1.5.2
pandocfilters==1.5.0
parso==0.8.3
pathy==0.10.0
pexpect==4.8.0
phonenumbers==8.13.1
pickleshare==0.7.5
platformdirs==2.5.4
pluggy==1.0.0
preshed==3.0.8
presidio-analyzer==2.2.30
presidio-anonymizer==2.2.30
presidio-evaluator==0.1.0
prometheus-client==0.15.0
prompt-toolkit==3.0.33
psutil==5.9.4
ptyprocess==0.7.0
pure-eval==0.2.2
pycountry==22.3.5
pycparser==2.21
pycryptodome==3.16.0
pydantic==1.10.2
Pygments==2.13.0
pyparsing==3.0.9
pyrsistent==0.19.2
pytest==7.2.0
python-crfsuite==0.9.8
python-dateutil==2.8.2
python-dotenv==0.21.0
pytz==2022.6
PyYAML==6.0
pyzmq==24.0.1
qtconsole==5.4.0
QtPy==2.3.0
regex==2022.10.31
requests==2.28.1
requests-file==1.5.1
schwifty==2022.9.0
scikit-learn==1.5.2
scipy==1.13.1
Send2Trash==1.8.0
six==1.16.0
sklearn-crfsuite==0.3.6
smart-open==5.2.1
sniffio==1.3.0
soupsieve==2.3.2.post1
spacy==3.4.3
spacy-legacy==3.0.10
spacy-loggers==1.0.3
srsly==2.4.5
stack-data==0.6.2
tabulate==0.9.0
terminado==0.17.0
thinc==8.1.5
threadpoolctl==3.5.0
tinycss2==1.2.1
tldextract==3.4.0
tomli==2.0.1
tornado==6.2
tqdm==4.64.1
traitlets==5.5.0
typer==0.7.0
typing_extensions==4.4.0
urllib3==1.26.13
wasabi==0.10.1
wcwidth==0.2.5
webencodings==0.5.1
websocket-client==1.4.2
widgetsnbextension==4.0.3
zipp==3.11.0
